alternates/matlab-installer/tests/requirements.txt
docker>=6
pytest-testinfra>=7